Wrapper of NARFCS enhanced forked mice version

This function is a wrapper of the NARFCS extension of the package `mice` by M. Moreno-Betancur, F. Leacy, D. Tompsett and I. White (https://raw.githack.com/moreno-betancur/NARFCS/master/Vignette.html) The function performs a sensitivity analysis based on the Not At Random Fully Conditional Specification (NARFCS) procedure. The wrapper intends to give users a more friendly version of the package, especially in situations with many covariates. The function itself follows the Formula syntax for fully conditionally specified models. For a great introduction to the overall NARFCS procedure and implementation, please visit <https://raw.githack.com/moreno-betancur/NARFCS/master/Vignette.html>

Usage

smdi_narfcs(
  i,
  data = NULL,
  id_var = NULL,
  covar = NULL,
  missing_indicator_var = NULL,
  predictor_covar = NULL,
  n_impute = 30
)

Arguments

i

numeric, sensitivity parameter delta; can also be a range of deltas if used in combination with lapply

data

dataframe or tibble object with partially observed/missing variables. Ideally consists only of continuous or binary covariates

id_var

character column name with (patient) identifier variable

covar

character covariate or covariate vector with variable/column name(s) to investigate; restricted to one covariate at this time which must be continuous or binary

missing_indicator_var

character indicating name of the missing indicator variable(s) for covar for narfcs (optional)

predictor_covar

character covariate vector for narfcs imputation procedure, if not specified takes all available variables in dataset

n_impute

integer, number of imputed datasets per delta parameter value

Value

returns a long mids object with multiply imputed datasets

Examples

if (FALSE) {

df %>%
smdi_narfcs()
}