Takes an object of class smdi and styles it to a publication-ready gt table

This function takes either an object of class smdi or data.frame or tibble as input and styles it to a publication-ready table based on the gt package. The output is of class gt and can take further gt-based arguments for customization.

Usage

smdi_style_gt(
  smdi_object = NULL,
  include_little = TRUE,
  font_size = 13,
  tbl_width = 800
)

Arguments

smdi_object

object of class "smdi" or data.frame/tibble

include_little

can be logical (TRUE/FALSE) for displaying Little's p-value that is part of an "smdi" object or a separate object of class "little"

font_size

integer to determine table font size

tbl_width

integer to determine table width

Value

returns a formatted gt table object

Details

See also

Examples

library(smdi)
library(dplyr)

smdi_diagnose(
  data = smdi_data,
  covar = "egfr_cat",
  model = "cox",
  form_lhs = "Surv(eventtime, status)"
  ) %>%
smdi_style_gt()

  

Covariate
      	ASMD (min/max)1
      	p Hotelling1
      	AUC2
      	beta univariate (95% CI)3
      	beta (95% CI)3
    

egfr_cat
	0.243 (0.010, 0.485)
	<.001
	0.629
	0.06 (95% CI -0.03, 0.15)
	-0.01 (95% CI -0.10, 0.09)

 p little: <.001,  Abbreviations: ASMD = Median absolute standardized mean difference across all covariates, AUC = Area under the curve, beta = beta coefficient, CI = Confidence interval, max = Maximum, min = Minimum
    

1 Group 1 diagnostic: Differences in patient characteristics between patients with and without covariate
    

2 Group 2 diagnostic: Ability to predict missingness
    

3 Group 3 diagnostic: Assessment if missingness is associated with the outcome (univariate, adjusted)